fix(seo): Refactor landing pages to Server Components for proper metadata
