+++ app/(landing)/loesungen/grosse-hausverwaltungen/page.tsx
@@ -1,159 +1,9 @@
-'use client';
+import { Metadata } from 'next'
+import { pageMetadata } from '@/lib/seo/metadata'
+import EnterprisePage from './content'
 
+export const metadata: Metadata = pageMetadata.loesungenGrosse
 
-import { TrendingUp, Database, Lock, Workflow, Globe, HeadphonesIcon, CheckCircle2, ArrowLeft } from 'lucide-react';
-import Link from 'next/link';
-import { Button } from '@/components/ui/button';
-import { useFeatureFlagRedirect } from '@/hooks/use-feature-flag-redirect';
-import { POSTHOG_FEATURE_FLAGS } from '@/lib/constants';
-
-export default function LargePropertyManagementPage() {
-  // Redirect to homepage if feature flag is disabled
-  const { isLoading, isAllowed } = useFeatureFlagRedirect(POSTHOG_FEATURE_FLAGS.SHOW_LOESUNGEN_DROPDOWN);
-
-  // Show nothing while checking feature flag to prevent flash of content
-  if (isLoading || !isAllowed) {
-    return null;
-  }
-
-  const features = [
-    {
-      icon: Database,
-      title: 'Enterprise-Skalierung',
-      description: 'Verwalten Sie unbegrenzt viele Wohnungen mit höchster Performance.',
-    },
-    {
-      icon: Workflow,
-      title: 'Custom Workflows',
-      description: 'Passen Sie Workflows an Ihre spezifischen Geschäftsprozesse an.',
-    },
-    {
-      icon: Lock,
-      title: 'Erweiterte Sicherheit',
-      description: 'SSO, 2FA und erweiterte Sicherheitsfeatures für Ihr Unternehmen.',
-    },
-    {
-      icon: Globe,
-      title: 'API & Integrationen',
-      description: 'Vollständige API-Zugriff und Integration mit Ihren bestehenden Systemen.',
-    },
-    {
-      icon: HeadphonesIcon,
-      title: 'Dedicated Support',
-      description: 'Persönlicher Account Manager und 24/7 Priority Support.',
-    },
-    {
-      icon: TrendingUp,
-      title: 'Advanced Analytics',
-      description: 'Umfassende Business Intelligence und Custom Reports.',
-    },
-  ];
-
-  const useCases = [
-    'Verwaltung von 100+ Wohnungen',
-    'Mehrere Standorte und Niederlassungen',
-    'Komplexe Organisationsstrukturen',
-    'Custom Branding und White-Label',
-    'Dedizierte Server-Infrastruktur',
-    'SLA-Garantien',
-    'Individuelle Schulungen für Teams',
-    'Maßgeschneiderte Entwicklungen',
-    'Compliance und Audit-Trails',
-    'Multi-Mandanten-Fähigkeit',
-  ];
-
-  return (
-    <div className="min-h-screen bg-gradient-to-b from-background to-muted/20">
-      <div className="container mx-auto px-4 py-16 mt-20">
-        <div className="max-w-6xl mx-auto">
-          {/* Hero Section */}
-          <div className="text-center mb-16">
-            <div className="inline-flex items-center justify-center w-20 h-20 bg-primary/10 rounded-2xl mb-6">
-              <TrendingUp className="w-10 h-10 text-primary" />
-            </div>
-            <h1 className="text-5xl font-bold mb-4">Für große Hausverwaltungen</h1>
-            <p className="text-xl text-muted-foreground max-w-2xl mx-auto">
-              Enterprise-Lösung für große Hausverwaltungen mit individuellen Anforderungen und höchsten Ansprüchen.
-            </p>
-          </div>
-
-          {/* Features Grid */}
-          <div className="mb-16">
-            <h2 className="text-3xl font-bold text-center mb-12">Enterprise Features</h2>
-            <div className="grid md:grid-cols-2 lg:grid-cols-3 gap-8">
-              {features.map((feature, index) => {
-                const Icon = feature.icon;
-                return (
-                  <div key={index} className="bg-card border rounded-2xl p-6 hover:shadow-lg transition-shadow">
-                    <div className="w-12 h-12 bg-primary/10 rounded-xl flex items-center justify-center mb-4">
-                      <Icon className="w-6 h-6 text-primary" />
-                    </div>
-                    <h3 className="text-xl font-semibold mb-2">{feature.title}</h3>
-                    <p className="text-muted-foreground">{feature.description}</p>
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-
-          {/* Use Cases */}
-          <div className="bg-card border rounded-2xl p-8 mb-16">
-            <h2 className="text-3xl font-bold mb-8 text-center">Enterprise-Funktionen</h2>
-            <div className="grid md:grid-cols-2 gap-4">
-              {useCases.map((useCase, index) => (
-                <div key={index} className="flex items-start space-x-3">
-                  <CheckCircle2 className="w-5 h-5 text-primary mt-0.5 shrink-0" />
-                  <span className="text-muted-foreground">{useCase}</span>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* Custom Solutions */}
-          <div className="bg-primary/5 rounded-2xl p-8 mb-16">
-            <h2 className="text-2xl font-bold mb-4">Individuelle Enterprise-Lösungen</h2>
-            <p className="text-muted-foreground mb-4">
-              Für große Hausverwaltungen bieten wir maßgeschneiderte Lösungen an. Kontaktieren Sie uns für
-              ein individuelles Angebot, das perfekt auf Ihre Bedürfnisse zugeschnitten ist.
-            </p>
-            <ul className="space-y-2 text-muted-foreground mb-6">
-              <li className="flex items-center space-x-2">
-                <CheckCircle2 className="w-4 h-4 text-primary shrink-0" />
-                <span>Individuelle Preisgestaltung basierend auf Ihren Anforderungen</span>
-              </li>
-              <li className="flex items-center space-x-2">
-                <CheckCircle2 className="w-4 h-4 text-primary shrink-0" />
-                <span>Persönliche Beratung und Implementierung</span>
-              </li>
-              <li className="flex items-center space-x-2">
-                <CheckCircle2 className="w-4 h-4 text-primary shrink-0" />
-                <span>Langfristige Partnerschaft und kontinuierliche Weiterentwicklung</span>
-              </li>
-            </ul>
-          </div>
-
-          {/* CTA Section */}
-          <div className="text-center bg-primary/5 rounded-2xl p-12">
-            <h2 className="text-3xl font-bold mb-4">Lassen Sie uns sprechen</h2>
-            <p className="text-muted-foreground mb-8 max-w-2xl mx-auto">
-              Vereinbaren Sie einen Termin mit unserem Enterprise-Team und erfahren Sie, wie wir Ihre
-              Hausverwaltung auf das nächste Level bringen können.
-            </p>
-            <div className="flex flex-col sm:flex-row gap-4 justify-center">
-              <Button asChild size="lg">
-                <Link href="/#cta">
-                  Beratungstermin vereinbaren
-                </Link>
-              </Button>
-              <Button asChild variant="outline" size="lg">
-                <Link href="/?getStarted=true">
-                  Demo anfordern
-                </Link>
-              </Button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
+export default function Page() {
+  return <EnterprisePage />
 }
